Implement delivery record management and performance reporting features

+++ src/javaassigmnet01/CostCalculator.java
@@ -4,14 +4,14 @@ public class CostCalculator {
     private static final double FUEL_PRICE = 310.0;  // LKR per liter
     private static final double MARKUP_RATE = 0.25;  // 25% markup
     
-    public static double calculateBaseCost(int distance, int ratePerKm, int weight) {
+    public static double calculateBaseCost(int distance, double ratePerKm, int weight) {
         // Cost = D × R × (1 + W/10000)
         return distance * ratePerKm * (1 + (weight / 10000.0));
     }
     
-    public static double calculateFuelUsed(int distance, int fuelEfficiency) {
+    public static double calculateFuelUsed(int distance, double fuelEfficiency) {
         // FuelUsed = D/E
-        return distance / (double)fuelEfficiency;
+        return distance / fuelEfficiency;
     }
     
     public static double calculateFuelCost(double fuelUsed) {
@@ -19,9 +19,9 @@ public class CostCalculator {
         return fuelUsed * FUEL_PRICE;
     }
     
-    public static double calculateDeliveryTime(int distance, int avgSpeed) {
+    public static double calculateDeliveryTime(int distance, double avgSpeed) {
         // Time = D/S
-        return distance / (double)avgSpeed;
+        return distance / avgSpeed;
     }
     
     public static double calculateOperationalCost(double deliveryCost, double fuelCost) {
